Move factory tests and rename service to integration tests
